dynamics.execute_repeatedly()
Improving the API of dynamics module.
More generic and flexible, and removing the need
to clutter it up with a long list of setters;
the task list *is* the setter.

Uses the arg-passing technique (see previous commit
976ffded) to allow tasks to require arbitrary
numbers of arguments. (Currently keyword; could
implement a way to accept positional later if
needed.)

Switched to new control method for tangent forces;
still need to implement for bond maintenance.