Comment/typing tweak: more consistent usage
Make comment consistent with the type-hinted
function declaration of execute_repeatedly()

In type hinting, intentionally did not use
"Optional", because caller should not say there is
"no" task list; there is always a task list! Even
if it can be empty. So, require caller to be
explicit about that, passing an actual list.

And that's even though the default arg is "None",
which is merely there because it's the proper
way to default to empty list.

Then also declare the private var consistent with
all that; it never needed to be Optional either.