Better solution for the Jupyter Notebook
See this stack overflow and all of its replies:
https://stackoverflow.com/questions/3319479/can-i-git-commit-a-file-and-ignore-its-content-changes

Two files now. The "template" one, checked in;
and the working version, which is a copy of the
template. The template is tracked by git, and the
working version is .gitignore-ed. Now always
run the working copy, never the template; that
way, changes won't be annoyingly tracked. And
anybody who gets the project via the repo will
receive only the template; they just have to copy
it. If this works as advertised, I will add
instructions to the file in a text cell.

(Note, the not-terribly-popular reply referring
to IntelliJ IDEA sounded intriguing, but upon
investigation I don't think it would work; I think
you'd have to always move changes into the
alternate changelist manually:
https://stackoverflow.com/a/33411930 )