sharon_utils external -> tf_utils internal module
Previously felt these utilities should be separate
from Epiboly project since they are intended to be
general purpose TF helpers, project-agnostic.
Now decided to bring it into the project after all.
Needed to bring it under version control, but
making a separate project for it didn't seem
justified. And changes are often coordinated with
changes in the main code.