Clean up tf_utils
1. Get rid of "vec". Originally it was a nice short
alias, "vec" instead of "mx.MxVector3f", saving
10 keystrokes. Then Tissue Forge made the name
shorter, "tf.fVector3"; meanwhile putting vec()
into a module made it longer, "su.vec", and now
longer still, "tfu.vec"; so now it only saves 4
keystrokes! Ditch it.

2. Move phi_for_epiboly into main. It never really
belonged here because it's project-specific; and
it's in fact only used in one place, so it's not
really general purpose even in the context of
the project. So move it to right where it's used.