Add global catalogs for particles, bonds
Map particle id to particle handle, because there's
no way to derive a handle from an id. (There's
supposed to be, but TF has an issue. Fix coming,
hopefully.) (And hopefully this works. If not,
may need to store actual Particles instead of
ParticleHandles.)

Map bond id to its potential's r0, because there's
no way to pull r0 from a potential (why not? should
work, but doesn't), and sometimes you can't even
get the potential from the bond.