Simplify global catalog
Now we know NOT to try to get Potential from an
inactive bond, which was the cause of the bug.
So, don't need to store potential in the catalog
after all. You can get it from the BondHandle.

However, still do need to store r0, because
apparently cannot get that from the Potential.
If you try, it always comes back as 0.0.