Comment on performance: wide-area neighbor-finding
Turns out it's not the sort that was causing the
performance problem. Experiments with benchmarking
showed the current algorithm (_attempt_closest_bond())
is just as fast as the much simpler previous one
(make_all_bonds()) except that it uses a wider area
neighbor search, which is what takes the time.