Iterative neighbor-finding with flexible definition of "nearby"
This removes (for all practical purposes) the
limitation of a maximum search distance, so a particle
to bond to can always be found.

At the same time, it is a performance improvement over
the non-iterative approach, because it only does slow,
long-distance searches when lack of available bonding
partners in the immediate vicinity makes it necessary;
when possible, searches are over short distances.

Despite these improvements, it did not solve the
problem I hoped it would solve, of the formation of
gaps. There's incremental improvement in that, too,
though: it does form bonds across the gaps, causing
stray particles to be "reeled in", in a way I didn't
see with the previous approach. This is a good sign,
and will at least contribute to the solution.