Refactor
Pass third particle to all the delta-energy terms.
For existing terms (neighbor count, and angle), they
don't need the particle, but did need to know whether
an edge transformation was happening, and they were
using other means to decide this (a flag, or checking
the particle types of the two particles whose bond
was making/breaking). Now just use this as a flag.
(Particle == the particle that will be changing
its type, if this is an edge transformation; or
"None" if it is not one.)

The new energy term (adhesion) needs the actual
particle (not yet implemented).