Take particle type transformation into account in adhesion energy (REALLY BAD IDEA!)
Notes:

1) I did this wrong. I only considered two of the bonds
on p_becoming: the two attached to the leading edge.
To correctly take the particle type transformation
into account, I need to consider ALL of the bonds
on that particle.

2) However, that's not worth doing, because this was
enough to convince me that this was a terrible idea.
Every particle wants to become internal, because
that automatically creates a favorable energy change!
So the entire leading edge goes away (at least until
the sim crashes due to numerical problems).