Start with 75 instead of 60 LeadingEdge particles
Previously, right at the beginning of the sim, about
20 particles would immediately become edge. They
would leap into the edge from way up in the 1st tier,
causing the margin to be quite uneven right off the
bat. This eliminated that, and the edge stays
straighter, longer. Overall result is same, but
initial dynamics are better.