Particle diffusion to prevent holes (experiment)
For now, setting up with k=0 at the start of the sim,
so that holes form; then can later enable diffusion
by increasing K, to see how much of a multiplier is
needed in order to heal the hole.

Guessing that less will be needed to prevent holes,
than to heal them. But this should be a quicker
way to get into the right ballpark.