Surprises and reminders!
Added Universe.time to image export filename, to
compare it with the "timestep" (the event invocation
count). Very surprising results, so I refactored the
usage of keyword arguments in tf.step() and
the equilibration function it is wrapped in, to
remind myself of the true meaning of "step"!

When called with no args, tf.step() indeed takes a
single step. Otherwise, the argument is not the
number of steps, it's an amount of time! I knew this,
but it's easy to forget.