Partially failed workaround: screenshots, no animation
Can now get screenshots during tf.step(), with a
small manual intervention up front. But all the
images look identical; they are not capturing the
simulation progress. Highly reminiscent of the
interactivity issues that were supposedly fixed in
TF v 0.0.2... which I cannot run because it's
broken in other ways.