Bug workaround: timer arithmetic
The real fix is to use on_event() instead of on_time(),
if you want an invoke on each timestep. But I could
not get that to work for some reason, so for now,
on_time() with period < dt ensures that each timestep,
enough sim time will have elapsed to generate an
invoke.