Stopping-point for free-running equilibration in windowless
When free-running equilibration in windowless mode,
set a short duration because there's no way to
gauge progress and no way to interrupt it manually
and still capture the MPL plot. This way it has a
natural end point.

(In windowed mode, duration is ignored since it will
run until the window is manually closed.)