Realistic values for number and size of cells
Number: adjusted to give a final number of cells
after all setup steps, approximately equal to the
number measured by Linlin (2252)

Size: once the number of cells was fixed in place,
use the tools created in the recent commits, to
determine the right radius to use, so that the
cells fit in the area above the leading edge.

Result: about 3x as many cells, at about half the
radius. Will probably need to make some other
adjustments as a consequence. Pay attention to the
sizes of forces, Potentials, etc.