Ensure minimum bond connectivity at initialization
Setup wasn't taking advantage of the iterative neighbor
search; it simply searched within a fixed area, and
sometimes bonded a particle to fewer than the minimum
number of neighbors. If that was 2, it was probably
non-ideal but went unnoticed (and probably got
remedied during bond maintenance). But when < 2
(more likely when experimenting with tuning the
cell radius and setup/equilibration process, when
larger gaps are present), it starts causing a crash
in operations that need at least two bonded neighbors.