Refactor troubleshooting code
So that it can be called from more places, and
break at the end of the sim for a final examination
without having to be triggered by an edge
transformation event.

Prediction: it will now always be caught at the
beginning of the sim; if it makes it past that first
check, then there will be no more problem for the
rest of the time.