memory-profiler: in the last state I used it:
As before, accept() will be called only once, whether the
transformation is accepted or not. But now ONLY in
make_bond, and not any of the others. And not profiling
any of the bond-rearranging functions, instead profiling
_move_toward_open_space(), but still using make_bond to
control how long it runs. So, will be one timestep,
one call to accept(), and also one call to
_move_toward_open_space() - which will hit every
particle once.

Experimental code in place: avoiding ParticleList.positions,
both in _move_toward_open_space and in
get_ordered_bonded_neighbors(). Profiling said this was
the culprit and changing it saved a lot of memory; but
when profiling the full script over time, that turned out
not to be the case. I do not understand why.