Treat config as static: no runtime changing
Refactor usage of screenshot_export_interval, which was
encouraging the use of config as a place to store
dynamically changing runtime values. Let module control
its own dynamic value internally; never change config at
runtime.

This will be important for saving/loading simulation state,
specifically the portion that's not handled by TF and I
have to handle myself. Now config can be assumed to be
unchanged, so does not have to be saved as part of
simulation state.