Protect sim state exports from keyboard interrupt
If a KeyboardInterrupt occurs during export, then data can be
corrupted or just missing entirely, and subsequent import
will fail. Did my best to prevent this by catching the
KeyboardInterrupt exception and repeating the export until it
succeeds; and only then exiting the app. Does not work for
^Z, only for ^C.

This was fraught with gotchas and took me an entire day
to get working.

This came up when TJ tried saving an export every single
timestep in order to expedite troubleshooting. But that
greatly increases the chances of aborting in the middle of
an export. Under normal circumstances, that should be much
less of an issue. (Unless, like for TJ, "normal circumstances"
means doing exactly that, in order to generate animations
during post-processing. There's gotta be a better way.)