Revert "Protect sim state exports from keyboard interrupt"
This reverts commit 4caacd3a8f8386029e0920ad28ce8ac54839706f.

This was a mess. Not really done correctly, and seemed to have some
problems. I'm probably better off without it. Just avoid aborting
the script in the middle of an export, which is usually easy,
assuming exports are far apart.

If I ever decide I need to be able to have exports very close
together (like every timestep) and also need to be able to
keyboard interrupt the script, then I'll need a better solution.
Maybe an on_keypress event, creating a clean and orderly exit.