Get r0 directly from potential (v0.0.2 improvement): UNTESTED
Was impossible in v0.0.1, so I had to store r0 in the
global catalog and retrieve it from there when needed.
Now remove that unneeded storage.

Reading r0 from potential should now work. But it's only
used in the bond relaxation feature which I am currently
not using, so it's UNTESTED. It compiles. Good enough!