Refactor names: it's tension, not bond length anymore
Because it's bond length relative to r0; the difference
is not bond length, rather it is a proxy for tension.
(Should be proportional to it.)

Also, a negative difference means the particles overlap;
that's not a valid tension. When overlapping, particles
are no longer subject to the bond potential at all (the
energy and the force should both be zero); whereas they
are subject to the type-based repulsion force. Just clip
that to zero and call that the tension.