Combo tension plot: replot the entire history each time
...instead of accumulating the plots over the course of
the sim. This means the history of all the plots is stored
globally as the actual data rather than as a plot, which is
probably the most straightforward way to then save its
state so that it can recover from a quit-and-restart
from saved state.

Also change the name of the _phi data to be more specific,
since the tension plots also use phi but in a different way.