Combo tension plot: Don't plot in reverse order after all.
Seems that fixing the upside-down legend was going to be
more trouble than it was worth. Also, doing it that way made
the color cycle shift each time the graph was re-plotted,
because the plots cycle through the colors in the order
they are plotted, and we inserted the newest plot at the
beginning of the sequence!

Instead, after plotting in chronological order, with the
T=0 line hidden at the back, "cheat" by plotting that
line again, in the same color in front.