Polar, circumf strain rates, WIP 2: bug fixes in calculation
(I found and fixed these bugs after implementing the new
plots, even though I'm committing the fixes first.)

1. I was getting NaN strain rates in cases where either
phi1 == phi2, or theta1 == theta2. This is because it
was trying to project onto a zero vector, which doesn't
work. Those vectors were getting zeroed out by the process
of accounting for the sign of the strain rate.

Solution: in those cases, the sign has to be non-negative
anyway, because compression isn't possible, only stretch.
So revert to the simpler calculation that doesn't account
for sign.

2. Operations on theta (averaging two of them; and taking
the difference between two of them) were incorrect because
they didn't take into account what happens at the boundary
between the range limits, i.e. when theta crosses over from
two pi to zero. Strangely, fixing this didn't seem to have
much effect on the appearance of the plots.