Plotting: Make time-averaging optional, with config flag.
Note that even when it's disabled, still leaving binsize at pi/20.
That will make PIV and speed-bin-diff plots a little more noisy,
but that's okay. I intend to leave it enabled most of the time.
Just needed to run it once more with disabled in order to get a
good comparison, since previous non-time-averaged plots were
prior to recent bug fixes.