Equilibration: Add final phase: balanced forces equilibration
Add yolk cortical tension (but not additional external force).
Add bond making and breaking.
This is like running the balanced-forces control, but only for
equilibration before proceeding into the simulation.

My hope was that this would make the T=0 lines on the various
aggregate plotted metrics look better, but it didn't really do
that, so not keeping this.

• Raw tension: values spread out a bit and moved up a bit; crossed
over the 0 line into negative, a little.
• Aggregate tension: moved up a bit, which was a slight
improvement when cell division disabled (because it stands out less
from the other lines, which are positive), but worse when cell
division enabled (because it stands out MORE from the other
lines, which are negative).
• PIV, and speed-bin-diffs: T=0 line just gets more jagged.
• Normal strain rate: just gets more jagged, which actually
makes it sand out less from the other jagged lines, so maybe
in this case it's a slight improvement?
• Polar and circumferential strain rates: just gets more jagged,
and goes to negative, which, as with aggregate tension, is
perhaps a little better when cell division disabled, but worse
when cell division enabled.

Tried this just with 10 time units (1000 timesteps) equilibration.
Didn't want to go longer since that already took 10 minutes. I
don't think that would really help anyway. This wasn't as helpful
as I'd hoped.