Equilibration: comment bit rot. Bug if cfg.show_equilibration == True
With all the recent enhancements and additions to plotting, there's
a crash. The very first time it tries to plot, there are no
bonds yet, so the plots that iterate over bonded pairs generate
no data points, and somewhere in one of the libraries it fails.
(I forget whether that was in matplotlib or numpy.)

This is because of the final equilibration phase, after the
leading edge z gets unfrozen, but before the bonds are added.
Would need to skip those particular plots during that phase.
Didn't bother for now, but will need to fix if ever need that
functionality again.