Adjustments after experimenting with dt and angle k
Original was dt = 0.01, k of angle bond = 1.0.
(Screenshot interval was 10 timesteps, plot interval was 2000.)

Changing only dt, I could go to 0.02 with no ill effects.
(Above that, the angle bonds blew up, or were jittery.)
(And screenshot interval was 5 timesteps, plot interval was 1000.)
(This was with uniform cell division, starting with the usual
applied force of 255, and approaching 1/10 of initial force as
circumference goes to 0.)

Provisional best result was by dropping k of angle bond to 0.5,
and nudging dt up to 0.025.
(Screenshot interval down to 4 timesteps, plot interval still 1000.)