Quick & Dirty(!!): config timers in Universe.time instead of time steps
So that when cfg.dt is adjusted, these timers don't have to be
adjusted to compensate. This is probably not the best way to do this,
but it's a quick and clean way to the result I need.

Note that I probably also want to change the plots to display
Universe.time (offset to the official sim start, after setup and
equilibration is finished), rather than the timestep count. But
use this for awhile, first, to make sure it's behaving well.