New bond-making algorithm (also new py 3.10 feature - match statement!)
Trying to fix holes in non-cell-division simulations. Seems to be
caused by reliance on bonding to nearest non-bonded neighbor. Try a
few other approaches.

Using BOUNDED with min_candidates = 1 and max_candidates = 7 seems
to help, though is not a complete solution. Seems to work about
half the time.

Was not helpful for cell-division sims, which seem better off without
it, so leave that as before: bonding to nearest non-bonded neighbor.

Also, match statement! New in python 3.10