Adjust "yolk cortical tension" for cell-division with variable force
Turns out that the balancing force needed adjustment in cell-division
sims, too, even though it's still using the same bond-making
algorithm as before. It's because of the non-constant force applied
to the leading edge; implemented back in June, and had not run
the control since before that (see June 15 run).

I speculate that having force depend on circumference creates a
feedback that makes the position of the leading edge much more of
an unstable equilibrium. With constant force, the little fluctuations
in position don't cause the force to change, so the balance is
stable; but with changing force, even a small fluctuation in
position triggers the feedback. So the balancing amount has to be
much more finely tuned, if nothing else.

So, contrary to the previous commit comment, this must also have
impacted the no-cell-division sims. But changing the bond-making
algorithm had an impact as well, because the new balancing force
depends on whether running with or without cell division. So
no-cell-division is affected by both.