Coupled bond remodeling, refactor #3: constraint energy calculations
Instead of passing one particle that will change (making or breaking),
pass two particles: one for making a bond, one for breaking a bond.
For now, not implementing the coupled case, just anticipating it.
Change function signatures to take both particles; and then these
functions no longer need to refer to the boolean "breaking" variable
from the enclosing function (accept()), because the status of the
two ParticleHandle values contains that information. We are making
a bond if making_particle is not None, and breaking a bond if
breaking_particle is not None. For now, one or the other of them
is guaranteed to be None.

neighbor-count constraint energy: easy. It will be zero in the
coupled case, so we can get this one over and done with right now.

bond-angle constraint energy: just fix the signature and leave
the calculation alone.