Coupled bond remodeling, refactor #4: filtering by constraints
In accept(), which decides whether a bond transformation is valid
and energetically permissible, instead of passing one particle that
will change (making or breaking), pass two particles: one for making
a bond, one for breaking a bond. For now, not implementing the
coupled case, just anticipating it. Change function signature to take
both particles; and eliminate the boolean "breaking" variable,
because the status of the two ParticleHandle values contains that
information. We are making a bond if making_particle is not None,
and breaking a bond if breaking_particle is not None. For now, one
or the other of them is guaranteed to be None.