Coupled bond remodeling, refactor #6 and implementation
Complete rewrite of the bond-angle constraint calculation,
delta_energy_angle(). It's the implementation of coupled bond
remodeling because it now can handle coupled events and compare
an existing bond with a potential new one and decide whether to
make the switch.

But, I haven't actually invoked it to do that yet. I want to
test it with the existing bond remodeling events first and make
sure I haven't broken anything. So it's also a refactor in the
sense that it's a giant no-op for now, invoking the same
remodeling events as always and evaluating them by the same
algorithm, with no expected change in behavior. Just using
different code that could also handle coupled events.

Also added definition of the coupled events themselves
(attempt_coupled_make_break_bond()), but it's not invoked.