Space-filling: make k_particle_diffusion a config var
Had been experimenting with k = 40, and it's fine when there's
no cell division going on (which is when I need this feature),
but when there's cell division it resulted in "popcorn" (particles
popping out of the layer and then settling back in). Force vector
is launching particle into space! But 20 seems to be plenty and
still solves the problem when cell division off.