Remove cruft: relax() function
Haven't used this in a long time, and won't be going back.
(viscosity = 0 meant it was disabled.) Also, missed a spot,
removing mention of space-filling from the comment, so this
is a good time to just get rid of the whole thing. Tagging
parent commit as the last one with that function in it.