Space-filling algo: improved for edge-bonded internal particles
Now go back to special treatment for ALL edge-bonded particles, not
just a special subset of them; but instead of disabling space-filling
entirely for those particles, use horizontal projection of the force
vector in order to get as much value from it as possible without
causing particle extrusion / ejection.

So, to summarize the result of recent tests and commits:

1. Space-filling modification: as described above.

2. Space-filling k=25 (higher values cause "popcorn" problem in
cell-division sim, which possibly could be fixed by accounting
for surface curvature when creating the force vector; that might
allow for higher k values. But at the moment, this seems good enough.)

3. Bondable-neighbor discovery: stick with the nearest neighbor and
don't bother with the more elaborate methods I had tried, for
exploring a larger set of particles. Not that helpful, and what
they achieved poorly, this new version of space-filling achieves
much better. And in fact the two don't play nicely together. (The
"BOUNDED" algorithm seemed to make the space-filling algorithm more
likely to result in particle extrusion, though I didn't analyze it
in depth. A whiff of it was enough.)