Minor clean-up
Previous commit moved this function from another module, including
the commented-out print() statement. But I added the final print(),
and used np instead of math since this module, unlike the one that
the function came from, already was importing np, but not math.
I just neglected to change the commented code, so doing that now.