Refactor get_ordered_bonded_neighbors() and _make_bond()
(In neighbors and bond_maintenance modules, respectively.)

Note: the point of this refactor was to enable me to reuse
get_ordered_bonded_neighbors() - originally used for finding bond
angles to which the energy-minimization constraint could be
applied - for more general purposes. Specifically, for "discovering"
which particles should be the EVL margin particles ("LeadingEdge"
at the time of this commit, though I'm planning on changing that
name to "EvlMargin"), instead of arbitrarily creating a ring of
them myself. It doesn't look like I'll be able to do that
successfully (I'll put the attempt in its own side branch for
posterity), but this is still a good refactor, so I'll keep it.
Breaks down the process into its parts, and exposes certain aspects
of it for use if needed.

1) Expose the calculated angles on which the list is sorted.

2) Instead of taking an arbitrary member of the list being sorted,
and using that as the reference angle, now use specifically the
vegetalward direction as the reference angle. This means the angles
returned will not be merely relative to one another, but instead
relative to an absolute heading on the embryo surface (vegetalward).
And this means that the most vegetalward angle on each side of
vegetal can be easily picked out (first and last items of the
returned list).

3) Previously, didn't bother sorting if there were < 4 items in the
list, because any order satisfied the description (the resulting
particle order should "trace a simple closed polygon around
particle p"). But now, have to sort those, to satisfy that we get
the result described in (2).

4) Also in bond_maintenance module, make function _make_bond()
visible (remove the underscore from the name) so it can be
called from elsewhere.