Attempt to discover leading edge (EVL margin) cells
I.e. find naturally which particles SHOULD be made leading edge,
rather than arbitrarily creating a ring of them. This did not
quite work. (Topology failures - it made a credible selection of
particles, but they were not connected properly, so the simulation
halted at one of my asserts.)

Don't think I can fix this, but don't want to toss it, so saving
it in its own branch. For history.