Cell division: Implement alternative particle splitting strategy
- Generate a vector tangent to the yolk surface at the particle,
randomly directed within the tangent plane, and pass that as the
direction vector to the split() overload; however, I discovered
this Tissue Forge feature is buggy! So, make it optional for now;
and added some debugging code which I will remove later.

- Double the volume of the particle (by multiplying radius by
cube root of 2), and its mass, prior to splitting, so that the
resulting daughters have the same value as before. Thus, no
more need to adjust size and mass afterward. Better for getting
TF to place the daughters properly next to each other.

- With these two changes, should need no post-splitting adjustment
to particle size, mass, or position, since the resulting
daughters should be the correct size and mass, and should be
placed exactly adjacent to one another. However, that's the
bug in the TF split() overload: they seem to be placed very far
apart!

- Also added some new RGB colors, to help with the debugging.