Movies: stop retaining the still image for the final movie frame
Retaining these to have the still image (higher resolution than
screenshotting later from the movie itself) has never been
terribly useful, because it turns out there's nothing special
about that final frame and it's no more likely to be the image
you want than any other time point. So you can't really predict
what frames you'll need.

And on the other hand, final_result_screenshots() was added
later, and does a much better job of capturing the final state
of the simulation, since it captures all 6 sides. (In contrast
to retaining the final frame, which in the context of epiboly
and camera rotation gets 4 rotational duplicates of the
vegetal pole!)

These images had useful information in their filename: a
record of the total number of cells at (or near) the end of the
simulation. Now that not retaining those files, put this in the
filename of the extra_state export file, instead.