Campinho data, hello world!
Num cells at 30% epiboly, and num divisions by 55% epiboly,
taken directly from their paper.

Particle radius selected to go along with that and just look good.
Large enough to be visible and render well, but small enough to
keep EVL appropriately thin, and to be distinctly less than
cell radius.

Worked right out of the box, thanks to all my preparation
leading up to this! There are some things to be tweaked but
this is very robust.

Note that the huge reduction in number of particles and number
of bonds makes this blazingly fast!