Change how mass is handled during a particle split (cell division)
TF wants to halve it, like volume (of its spherical particles).
I had been suppressing that, making it stay the same, like my
particle size did. But now that particle size is mostly irrelevant
to the physics, and it's cell size that matters, need to treat
mass accordingly. I believe that in Overdamped dynamics, it
should go down like surface area does, not like volume does.
And for squamous cells, constrained to move in a sheet, I
believe the relevant surface area is the lateral cell surface,
excluding apical and basal; thus for this model, that means it
should decrease like circumference does, hence like radius
does (linear dimensions). So net change, mass *= sqrt(2)/2.

Along with this, drop dt down to 0.1. It was necessary when
experimenting with letting mass be halved; that blew up at
dt = 0.16. With mass *= sqrt(2)/2, 0.16 did not blow up, but
clearly we are near the edge. Safer to stick with 0.1.