Fix initial cell radius calculation; fixes negative tensions
Negative tensions were because my cell radius calculation was off.
Incorporate circumscribed hexagon into calculation of cell
footprint; i.e., treat the calculated footprint (total EVL area /
number of cells) as a hexagon, not a circle - so that it includes
the "spandrels"! Then base cell radius on the inscribed circle,
NOT the radius of a presumed circular footprint. This results in
a radius about 4.5-5% lower. Cells before were slightly too big,
hence crowded. Now tension at T=0 is 0 again, as it should be.