Update external force, bond spring constants, and space-filling k
I've changed a lot of things lately, and have had to adjust
these constants as I go. These new values have been working well
for awhile, so this is a good time to check them in. I'm
probably not done changing them, though. They are pretty sensitive
to other changes in the model.