Recoil test, with new cell division
Now works fine even when cell division turned on.
There's plenty of room to contract without becoming solidly
packed. I assumed it was because of the difference in cell and
particle size, or because of the size change at cell
division, but no. It's actually because cell division ceases at
55% epiboly. Thereafter, cells become stretched as the tissue
stretches. That's why there's plenty of relaxation space; by
the end of epiboly, the cell division case resembles the no-
cell-division case: stretched cells, spaced well apart. So,
I never tried it back then, but I think as soon as I introduced
that division cessation prior to end of epiboly, even with
the old uniform cell sizes, this would have worked already.