Global catalog: capture bond spring constant for later retrieval
We want to calculate tension on a bond correctly; now that
different bonds have different spring constants, we need to take
those into account in the tension calculation for any given
bond. But, we can't read that value out of the bond's potential
(TF doesn't make it accessible), so we need to capture it and
then be able to retrieve it later.

(Note, have been using this for awhile, so may be logically prior
to a bunch of different commits. i.e. a bunch of the simulation
runs dated prior to this commit, have already been using this.)