Fix broken strain/stress calculations
These are used only in plotting, and they assumed constant r0
and spring constants.

r0 is now variable, reflecting cell size before/after division,
so the strain calculation now needs that denominator in it,
to be correct.
(Also, minor refactor on this: get strain on an individual
bond, or the aggregate strain of all bonds on a particle.)

Furthermore, strain can no longer be used as a proxy for
tension, because that's only valid when all spring constants
are the same, and we now have varying spring constants
(when cfg.spring_constant_cell_size_factor != 1). So, take
spring constant into account when calculating tension.

(Note: strain lives in tf_utils, generic for any Tissue Forge
use, but tension has to live in epiboly_utils, depending on
this simulation code, because it depends on global_catalog
in order to retrieve the spring constants, which are not
accessible from tf.Potential objects.)

(Note, have been using this for awhile, so may be logically prior
to a bunch of different commits. i.e. a bunch of the simulation
runs dated prior to this commit, have already been using this.)