Global catalogs: bugfix -> (but new bug)
Forgot that now bonds_by_id has real data again, so it has to be
stored as part of state. Also update all the commentary to match.

While at it, change the key names for both bond and particle
dictionaries, to save space on disk. Ridiculous to save thousands
of copies of the same word.

However, now It's bombing on re-import. Program is searching
in the dictionary for bond ids that are not there. (Both to del
them when bond is broken, and to retrieve spring constants.)
Turns out this is because the bond ids are not preserved after
re-import! Committing this anyway, because for now, re-import
bombs whether I have this or not. Clearly I need this code.
I have filed an issue:
https://github.com/tissue-forge/tissue-forge/issues/65