Bond remodeling: edge target angles base on polygon
Instead of 180°, let the target angle take the curvature into
account. We already had the calculation for that, left over from
the Angle bond days; now put it to better use.

While at it, make the function name more general. ("harmonic
angle" referred to the special TF Potential type to be used
with Angle bonds.)