Force algorithm: refactor / reframe
Improve variable naming, for better self-documentation (first
loop).

Importantly, reframe the way force is apportioned to particles,
so that it makes more sense when I write about it (second loop);
remove reference to global force (total force on entire
circumference) because it should be a local calculation.

The calculation doesn't actually change at all; it just has its
parentheses moved around. Multiplying three items:
force-to-circumference ratio, circumference, and particle weight.

Previous framing was to multiply the first two as an intermediate
result (that ratio, times the circumference): we're calculating
the total global force, then apportioning that among the
particles according to their weights.

Better framing is to take the LAST two as the intermediate result
(circumference, times particle weight): we're apportioning the
circumference among the particles, i.e., how much of the
circumference is controlled by each particle, or how much is each
particle "responsible for"? That's still a reference to a global
(circumference), but it's not global *force*, it's merely
geometry. Once we know the weighted or effective cell width,
we can multiply it by that ratio to give the local amount of
force. We've calculated the forces on each particle with only
local references.