Force algorithm: refactor / reframe (better & better variable names)
And now that I've written the description of this in the ms.,
finding the best way to describe it, change the variable names
once more and add an additional intermediate named result, to
match exactly the way I describe it in the ms.

Mainly, ditch the literal descriptive "arc_to_neighbors" and
"arc_to_veg_pole", to name them instead for what they
represent: relative cell width and relative an-veg distance.
Also name the normalized weights.