Preventing runaway edge recruitment - bugfix
Turns out there was a bug, and it was overcorrecting, and
that's why there's been NO edge recruitment lately. The bug
dates back to when I first switched from using z to using phi
as the criterion, but only got severely exposed after recent
changes to cell size and cell division.

After fixing the bug, there was lots of runaway edge
recruitment, i.e. it was now undercorrecting, so had to make
the threshold actually more stringent.