Bond maintenance: don't let margin ring get below 4 particles
Should never happen since we stop the main sim before it
ever gets down that far. But there is a bug in the recoil test,
where it "recoils" in the wrong direction and the margin
shrinks all the way down. It gets down to 3 particles in the ring
and then any attempt to remove one more particle would violate the
connectivity rules and hit an assert. Fix is to reject the
illegal operation until I can figure out what's up with the
recoil test.