Bond maintenance: min margin length = 4 to prevent assert
This only matters during recoil test, because otherwise the
margin never gets that small. If a particle is allowed to
leave, bringing the margin down to only 3 cells, for some
reason the geometry often gets messed up and you
end up with an edge particle with only one bonded edge
neighbor. I have no idea why, so just prevent it.

(And if that doesn't happen, and another particle leaves,
bringing the margin down to 2 cells, then you always assert,
because it involves bonding two particle that are already bonded.)