Plotting: Fix inappropriate interdependency (refactor)
_timestep was being appended to _timesteps list by the
progress_graph, which was the first plot I ever implemented.
But now most of the other plots use it as well. It worked
because that is also the first plot executed, but that means
the other plots had an inappropriate dependency on the order
of execution of the various plot functions. So move this to
the right place.